feat: add toggle button
