+++ src/App.js
@@ -1,7 +1,7 @@
 import { Route } from 'react-router-dom'
 import './App.css'
 import About from './pages/About/About'
-import Home from './pages/Home'
+import Home from './pages/Home/Home'
 import Footer from './presentational/Footer/Footer'
 import Header from './presentational/Header/Header'
