+++ src/container/SettingsContainer/SettingsContainer.js
@@ -1,9 +1,11 @@
 import React from 'react'
+import SetSoundNoti from '../../presentational/SetSoundNoti/SetSoundNoti'
 import './SettingsContainer.css'
 
 const SettingsContainer = () => {
     return (
         <div id="settings-container">
+            <SetSoundNoti/>
         </div>
     )
 }
