+++ src/pages/Home/Home.css
@@ -1,0 +1,3 @@
+div#home {
+    min-height: 70vh;
+}
