+++ src/presentational/SetSoundNoti/SetSoundNoti.css
@@ -1,0 +1,46 @@
+@import url('../../App.css');
+
+.label {
+    display: inline-flex;
+    align-items: center;
+    color: #394a56;
+    font-size: calc(10px + 1.2vw);
+    font-family: "nanum-square-round";
+}
+  
+.toggle {
+    isolation: isolate;
+    position: relative;
+    width: 5vmax;
+    height: 2.5vmax;
+    border-radius: 1.2vmax;
+    margin-left: 2vw;
+    overflow: hidden;
+    box-shadow:
+      -8px -4px 8px 0px #ffffff,
+      8px 4px 12px 0px #d1d9e6,
+      4px 4px 4px 0px #d1d9e6 inset,
+      -4px -4px 4px 0px #ffffff inset;
+    cursor: pointer;
+}
+  
+.toggle-state {
+    display: none;
+}
+  
+.indicator {
+    height: 100%;
+    width: 200%;
+    background: #ecf0f3;
+    border-radius: 15px;
+    transform: translate3d(-75%, 0, 0);
+    transition: transform 0.4s cubic-bezier(0.85, 0.05, 0.18, 1.35);
+    box-shadow:
+      -8px -4px 8px 0px #ffffff,
+      8px 4px 12px 0px #d1d9e6;
+}
+  
+.toggle-state:checked ~ .indicator {
+    transform: translate3d(25%, 0, 0);
+    background-color: var(--sub-color);
+}
